### 2023-11-20 #6068 @kyleniemeyer
Hello @samwaseda, thanks for your submission to JOSS. Unfortunately, we do not have an editor available to handle this right now, so I have to place this on our waitlist until someone becomes available (hopefully within a week or two).

In the meantime, please take a look at those missing DOI messages in your paper.
