+++ paper.bib
@@ -12,6 +12,7 @@
   number={10},
   pages={103202},
   year={2014},
+  doi={10.1088/0953-8984/26/10/103202},
   publisher={IOP Publishing}
 }
 
@@ -22,6 +23,7 @@
   volume={264},
   pages={107938},
   year={2021},
+  doi={10.1016/j.cpc.2021.107938},
   publisher={Elsevier}
 }
 
@@ -42,6 +44,7 @@
   volume={7},
   number={1},
   pages={103--109},
+  doi={10.1051/jphystap:019170070010300},
   year={1917}
 }
 
@@ -80,6 +83,7 @@
   journal={Proceedings of the National Academy of Sciences},
   volume={97},
   number={18},
+  doi={10.1073/pnas.97.18.9840},
   pages={9840--9843},
   year={2000},
   publisher={National Acad Sciences}
@@ -103,6 +107,7 @@
   volume={163},
   pages={24--36},
   year={2019},
+  doi={10.1016/j.commatsci.2018.07.043},
   publisher={Elsevier}
 }
 
