### 2024-01-09 #6068 @vipinagrawal25
Yes.  I can do that.

Regards
Vipin

> On 09-Jan-2024, at 12:23 AM, Kelly L. Rowland ***@***.***> wrote:
>
>
> hi @vipinagrawal25 <https://github.com/vipinagrawal25> 👋 would you be interested in and available for reviewing this JOSS submission?
>
