### 2024-01-09 #6194 @editorialbot
[REVIEW]: mamonca - interactive magnetic Monte Carlo code
**Submitting author:** @samwaseda (<a href="http://orcid.org/0000-0002-1677-4057">Osamu Waseda</a>)
**Repository:** https://github.com/samwaseda/mamonca
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.0.8
