### 2024-01-22 #6194 @arjunsavel
Thanks to @samwaseda for submitting an interesting package! `mamonca` is a modular code for simulating magnetic interactions that fits in neatly with existing codes from the literature.

My comments are below — once they are addressed, I recommend this for publication.

### Documentation notes
- The documentation could benefit from a short description of how to run the tests (I successfully ran them).
- The documentation could also benefit from a more detailed statement of need. I.e., why compute Metropolis Monte Carlo simulations?
-  I couldn't verify the `conda` installation (my os is arm64). Perhaps mention that only linux installation is currently supported for the `conda` install?
- The `cython` dependency is not currently listed — setup.py fails without a `cython` install.
- The C code of `mamonca` is largely uncommented. I understand that the docstrings are in the Python interface, but commenting through the C logic would make it more readable.
- The package does not currently contain contributing guidelines. For an example, see the [Numba contributing guidelines](https://github.com/numba/numba/blob/main/CONTRIBUTING.md).
- Pyiron is technically a dependency for running the provided examples — on that basis, I think that it should be added as a requirement.
- Once we run a simulation, what are some of the properties we can `get`? This information is in the docstring(s), and it should also be in the Quickstart.
- After executing the `set_repeat` method, I get the following error: `DeprecationWarning: pyiron_atomistics.atomistics.structure.atoms.set_repeat is deprecated: Use Atoms.repeat. when running structure.set_repeat`. Could this error be suppressed, circumvented, or explained?
- In the "First Steps" in the README, each line should be explained in more depth. E.g., what does `.set_repeat(10)` do? Why `10`? What do I do with the output? How do I know it’s run successfully? What things can go wrong?
- When `gett`ing the mean energy, are these units also in eV? Same question about the energy variance.
- The README could point toward the "First Steps" notebook for users interested in further examples.

### Notebook Notes
- The Curie temperature is noted as being “higher than it is supposed to be”. What is it supposed to be?
- I got slightly different plots when I ran the notebook cells — could the authors emphasize somewhere in the documentation (ideally the README) that the Monte Carlo nature of the calculation will induce some random error between calculations? Is there a recommendation to run the code a number of times? By how much should I increase the number of iterations?
- What parameters(s) did we choose randomly?
- Just checking that the magnetization is indeed unitless.
- As in the README, more commenting throughout would be helpful. For instance, what's happening in the loop over `ti_lambda`? What is fcc?
- Ideally, under each heading, there’d be a description of what the goal of that subsection is.
- What is the `lambda` being evaluated?

### Paper notes
- Line 7: Ranging from iron to what? The authors should include another example (e.g., "ranging from iron to copper") or leave iron as the only example (e.g., "for instance, iron")
- Line 11: "...compute the potential energy" the authors could provide an example of what systems the potential energy is computed for — even if it's just a generic lattice.
- Usually, the package itself and its features are described briefly in the Summary.
- A brief description of the differences between the Monte Carlo method and the spin dynamics method would be helpful (especially if there are any obvious pros/cons to each).
- Line 21: "on-the-fly" should not be hyphenated in this usage (because it doesn't precede the thing it's modifying)
- The authors could describe the workflow with `pyiron` a bit more — e.g., what input parameters for the Hamiltonian are provided.
- Are there specific benefits to running both metadynamics and thermodynamic integration in one code?
- It might be helpful to include a plot showing typical output from `mamonca`.
