### 2024-02-23 39f0e5c
add pyiron installation
