### 2024-02-23 10fc67a
add indication to first_steps.ipynb
