+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+# Contributing to mamonca
+
+Welcome to mamonca! We appreciate your interest in contributing.
+
+## How to Contribute
+
+1. Fork the repository to your GitHub account.
+2. Clone the forked repository to your local machine:
+
+   ```bash
+   git clone https://github.com/samwaseda/mamonca.git
+   ```
+
+Feel free to submit any changes.
