### 2024-02-27 #6194 @arjunsavel
Thanks to @samwaseda for addressing my comments! The changes you made have made the package even cleaner.

I have a few (hopefully small) final requests:
- in the paper: "while spin dynamics would deliver also the kinetics —> while spin dynamics also delivers the kinetics"
- What are the units (if any) for "lambda" in the first steps notebook?
- What do you think of adding pyiron to the setup.py requirements?
- Could follow one of the suggestions [here](https://stackoverflow.com/questions/37471313/setup-requires-with-cython) so that Cython isn't imported in setup.py before it's installed

Also, the LaTeX in the readme is a nice touch!
