### 2024-02-27 #6194 @samwaseda
Thanks again for the comments! Now I made changes accordingly!

> * What are the units (if any) for "lambda" in the first steps notebook?

I added "unitless" in the notebook

> * What do you think of adding pyiron to the setup.py requirements?

That's in principle a great idea (and sorry for not commenting on this in my last entry). One main problem is that it takes a very long time to install pyiron, so those who don't need pyiron would definitely not want to install pyiron for no reason. I added the comments in README saying how to install pyiron, so I'm hoping it will be enough. What do you think?

> * Could follow one of the suggestions [here](https://stackoverflow.com/questions/37471313/setup-requires-with-cython) so that Cython isn't imported in setup.py before it's installed

Thanks for your support! I didn't know much about this part, but now I changed it according to what's written there. I hope now it works!
