### 2024-02-27 #6194 @arjunsavel
Thanks for your quick response!

> That's in principle a great idea (and sorry for not commenting on this in my last entry). One main problem is that it takes a very long time to install pyiron, so those who don't need pyiron would definitely not want to install pyiron for no reason. I added the comments in README saying how to install pyiron, so I'm hoping it will be enough. What do you think?

This makes sense. Thank you for the explanation!

> Thanks for your support! I didn't know much about this part, but now I changed it according to what's written there. I hope now it works!

Thanks for working on this! I think the last thing to do, in addition to the setuptools version pin, is something like the below:

```
from setuptools.command.build_ext import build_ext
from setuptools import setup, Extension

ext = Extension(
    'mamonca',
    sources=["mamonca/mc.pyx"],
    language="c++",
    extra_compile_args=['-std=c++11'],
)

with open('README.md') as readme_file:
    readme = readme_file.read()

setup(
    name='mamonca',
    version='0.0.8',
    description='mamonca - interactive Magnetic Monte Carlo code',
    long_description=readme,
    long_description_content_type='text/markdown',
    url='https://github.com/samwaseda/mamonca',
    author='Sam Waseda',
    author_email='waseda@mpie.de',
    license='BSD',
    cmdclass={"build_ext": build_ext},
    ext_modules=[ext],
    options={'build': {'build_lib': 'mamonca'}},
    setup_requires=[
        # Setuptools 18.0 properly handles Cython extensions.
        'setuptools>=18.0',
        'cython',
        'numpy',
    ],
)
```

This way, Cython never needs to be imported at the top of the file, but `setuptools` is told that it requires Cyython to build.
