### 2024-02-27 #6194 @vipinagrawal25
Dear @samwaseda,

Firstly, please accept my apologies for the delay in completing the review of your paper. Thanks for the submission. I also would like to thank the editor @kellyrowland for choosing me as a reviewer and @arjunsavel for the thorough review.

The `mamonca` package is built for computing free-energy for materials using Heisenberg Landau model. Additionally the package also offers magnetic thermodynamic integration and metadynamics. I think the community will be benefitted from such software hence I recommend this for publication provided the comments outlined below are addressed.

**General:**
- I have observed that most commits are made by @samwaseda, followed by [Simon Bekemeier](https://github.com/s4b7r) with 4 commits, and [Jan Janssen](https://github.com/jan-janssen) with one commit. However the paper does not include these names in the co-author list but two other names which have not contributed to the repository. In case, if these co-authors have made non-code contributions to the work, I recommend including an author contribution section. Also note that, the [Joss guidelines] (https://joss.readthedocs.io/en/latest/submitting.html#authorship) emphasise that,
> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.

** Documentation**
- It would be beneficial to include information about dependencies in the README file. For instance, user should install Cython before using the package.
- While the `mc.pyx` file contain good comments, I suggest adding comments to `cMC.cpp` as well. Additionally, the use of [Doxygen](https://www.doxygen.nl) could benefit the user.

**Examples**
- I was unable to locate any examples on spin-dynamics. Can you please consider adding one example on that as well?
- While the tests provided by authors are running as expected, additional information is needed to verify the correctness of results. The authors could include ways to benchmark the code. For instance:
  - Compute statistical quantities to validate the results from MC simulation.
  - Reproduce well-documented results.
  - Compare results with those obtained from other existing codes.
  - Similar benchmarks should be considered for magnetic thermodynamics integration and Metadynamics.

**Paper**
- It would be useful to include a statement about the target community. Are the intended users experimental material scientists running small simulations, or individuals involved in modeling?
- Regarding the reference for understanding magnetic thermodynamic integration by Frenkel & Smit, please cite the specific chapter that explains the relevant part.
- The paper could be enhanced by including a section on explaining the Heisenberg-Landau model, detailing how MC methods and spin dynamics are employed for such systems. Adding a figure would be helpful for better understanding. @samwaseda has already given explanation about not including figures related to thermodynamic integration or Metadynamics but they can consider putting up the plots for the former part.
