### 2024-05-14 #6194 @vipinagrawal25
@samwaseda I realise that it can be a lot of work to do that. I had mentioned four points to verify the correctness of results.

1) Compute statistical quantities to validate the results from MC simulation.
2) Reproduce well-documented results.
3) Compare results with those obtained from other existing codes.
4) Similar benchmarks should be considered for magnetic thermodynamics integration and Metadynamics.

Can you please tell me which of these steps have been completed and which are pending?
