### 2024-06-14 #6194 @samwaseda
I have a question: The points raised by @vipinagrawal25 (comparison with existing results etc.) should be inside the paper, or is it enough to refer to notebooks in the same repository?
