+++ .github/workflows/testing.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.7, 3.9, '3.10']
+        python-version: [3.10, 3.11, 3.12]
 
     runs-on: ${{ matrix.os }}
     steps:
