+++ .github/workflows/testing.yml
@@ -16,16 +16,16 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v2
-    - name: Cache conda
-      uses: actions/cache@v1
-      env:
-        # Increase this value to reset cache if etc/example-environment.yml has not changed
-        CACHE_NUMBER: 0
-      with:
-        path: ~/conda_pkgs_dir
-        key:
-          ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{
-          hashFiles('.ci_support/environment.yml') }}
+#     - name: Cache conda
+#       uses: actions/cache@v1
+#       env:
+#         # Increase this value to reset cache if etc/example-environment.yml has not changed
+#         CACHE_NUMBER: 0
+#       with:
+#         path: ~/conda_pkgs_dir
+#         key:
+#           ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{
+#           hashFiles('.ci_support/environment.yml') }}
     - uses: conda-incubator/setup-miniconda@v2
       with:
         miniforge-variant: Mambaforge
