### 2024-06-14 e485c78
re-introduce 3.12
