### 2024-06-17 #6194 @samwaseda
Sorry for going back and forth, but it looks to me like other articles (including the example article on the JOSS page) tend to summarise the functionalities of the code but not show comparison with other codes. Should I really include figures showing a comparison between my code and other results? Otherwise I already included a jupyter notebook showing such a comparison.
