### 2024-06-17 #6194 @kellyrowland
Inclusion of the notebook is fine; you can refer the reader to that.
