### 2024-06-17 #6194 @samwaseda
Ok in that case I now updated the document indicating the notebooks. Please let me know if the format is good enough.

> 1. Compute statistical quantities to validate the results from MC simulation.
> 2. Reproduce well-documented results.
> 3. Compare results with those obtained from other existing codes.
> 4. Similar benchmarks should be considered for magnetic thermodynamics integration and Metadynamics.

I think now 1, 2 and 3 are fulfilled. For 4, it’s done for Metadynamics, but it’s difficult to prove it for thermodynamic integration. I included an example but it’s not a benchmark. I hope it’s still acceptable.
