### 2024-06-18 7fb1485
decrease miniconda version
