+++ .github/workflows/testing.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
-    - uses: conda-incubator/setup-miniconda@v4
+    - uses: conda-incubator/setup-miniconda@v2
       with:
         miniforge-variant: Mambaforge
         activate-environment: mamonca-test
