### 2024-06-18 e842ab5
change channel priority to flexible
