### 2024-06-18 6fb1405
bring back channel-priority
