+++ .github/workflows/testing.yml
@@ -20,7 +20,7 @@ jobs:
       with:
         miniforge-variant: Mambaforge
         activate-environment: mamonca-test
-        channel-priority: flexible
+        channel-priority: strict
         environment-file: .ci_support/environment.yml
         python-version: ${{ matrix.python-version }}
         use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
