+++ .github/workflows/testing.yml
@@ -23,7 +23,6 @@ jobs:
         channel-priority: strict
         environment-file: .ci_support/environment.yml
         python-version: ${{ matrix.python-version }}
-        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
 
 
     - name: run tests
