+++ .github/workflows/testing.yml
@@ -19,15 +19,16 @@ jobs:
     - uses: conda-incubator/setup-miniconda@v2
       with:
         miniforge-variant: Mambaforge
-        activate-environment: mamonca-test
+        python-version: ${{ matrix.python-version }}
         channel-priority: strict
         environment-file: .ci_support/environment.yml
-        python-version: ${{ matrix.python-version }}
-        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
 
+    - name: Install dependencies
+      run: conda env update --file .ci_support/environment.yml --name base
 
     - name: run tests
       shell: bash -l {0}
       run: |
         pip install -e .
         python -m unittest discover tests/
+
