+++ .github/workflows/testing.yml
@@ -16,14 +16,14 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v4
+    - uses: conda-incubator/setup-miniconda@v2
       with:
+        miniforge-variant: Mambaforge
+        activate-environment: mamonca-test
+        channel-priority: strict
+        environment-file: .ci_support/environment.yml
         python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies
-      run: |
-        pip install numpy cython scipy
+        use-only-tar-bz2: true
 
     - name: run tests
       shell: bash -l {0}
