### 2024-06-18 2c591d3
remove python 3.12
