### 2024-08-12 #6194 @kellyrowland
@samwaseda I'm starting the wrap-up process for this since the reviews have been completed. It looks like you'll want to add the following DOIs to the references:

- `10.1088/1742-5468/2011/05/P05001` for `The ALPS project release 2.0: open source software for strongly correlated systems`
- `10.1103/PhysRevB.63.052405` for `Ab initio calculation of phase boundaries in iron along the bcc-fcc transformation path and magnetism of iron overlayers`
- `10.1016/j.cpc.2017.04.006` for `Quantum lattice model solver HΦ`
- `10.1088/1361-648X/aa680e` for `The atomic simulation environment-a Python library for working with atoms`
- `10.1103/PhysRevB.103.024421` for `Ab initio based models for temperature-dependent magnetochemical interplay in bcc Fe-Mn alloys`

(the DOIs may need to be reformatted; I just grabbed them from the publications' webpages)
