+++ paper.md
@@ -13,19 +13,17 @@ authors:
     equal-contrib: true
     affiliation: 1
   - name: Tilmann Hickel
+    orcid: 0000-0003-0698-4891
     affiliation: 1
   - name: Jörg Neugebauer
     affiliation: 1
+    orcid: 0000-0002-7903-2472
 affiliations:
  - name: Max-Planck-Institut für Eisenforschung, Max-Planck-Straße 1, D-40237 Düsseldorf, Germany
    index: 1
 date: 1 November 2023
 bibliography: paper.bib
 
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
