+++ paper.bib
@@ -34,6 +34,7 @@
   volume={217},
   pages={180--192},
   year={2017},
+  doi={10.1016/j.cpc.2017.04.006},
   publisher={Elsevier}
 }
 
@@ -56,6 +57,7 @@
   number={5},
   pages={052405},
   year={2001},
+  doi={10.1103/PhysRevB.63.052405},
   publisher={APS}
 }
 
@@ -67,6 +69,7 @@
   number={05},
   pages={P05001},
   year={2011},
+  doi={10.1088/1742-5468/2011/05/P05001},
   publisher={IOP Publishing}
 }
 
@@ -97,6 +100,7 @@
   number={27},
   pages={273002},
   year={2017},
+  doi={10.1088/1361-648X/aa680e},
   publisher={IOP Publishing}
 }
 
@@ -119,5 +123,6 @@
   number={2},
   pages={024421},
   year={2021},
+  doi={10.1103/PhysRevB.103.024421},
   publisher={APS}
 }
