### 2024-08-13 #6194 @samwaseda
> @samwaseda I'm not seeing any updates to https://github.com/samwaseda/mamonca/blob/joss-paper/paper.bib with the DOIs - maybe a commit hasn't been pushed?

Sorry you are right. I just pushed the changes!
